gcp/edge-to-cloud-video-analytics-media-cdn-vertex-ai/code/terraform/function-templates/requirements.txt
# Cloud Functions requirements for Video Analytics Pipeline
# Python 3.9+ compatible packages

# Google Cloud SDK dependencies
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
google-cloud-functions==1.8.3
google-cloud-videointelligence==2.11.3
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0

# Core Python packages
requests>=2.31.0
urllib3>=1.26.18

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0

# JSON and data handling
jsonschema>=4.17.0

# Date and time utilities
python-dateutil>=2.8.2

# Async support
asyncio>=3.4.3

# Function framework
functions-framework>=3.4.0

# Optional ML and processing libraries
Pillow>=10.0.0  # For image processing if needed
opencv-python-headless>=4.8.0  # For video frame analysis if needed

# Performance and utilities
cachetools>=5.3.0
retrying>=1.3.4

# Security and validation
cryptography>=41.0.0

# Development and debugging (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0